Rename spacetime modules and update imports
Renamed L0_granule_based.py to L0_granule_grid.py and L1_field_based.py to L1_field_grid.py for clarity and consistency. Updated all relevant imports, documentation, and references throughout the codebase. Also added PI and EULER_NUMBER constants in openwave/common/constants.py and made minor doc/comment improvements.
