+++ openwave/common/constants.py
@@ -52,7 +52,7 @@ EWAVE_PERIOD = 9.520241169e-26  # s, energy-wave period (T = 1 / EWAVE_FREQUENCY
 # Neutrino particle (seed particle)
 # ================================================================
 NEUTRINO_K = 1  # neutrino wave center count (dimensionless)
-NEUTRINO_RADIUS = EWAVE_LENGTH  # m, neutrino radius = 1 λ
+NEUTRINO_RADIUS = EWAVE_LENGTH  # 2.85e-17m, neutrino radius = 1 λ
 NEUTRINO_ENERGY = 3.8280e-19  # J, neutrino "seed" energy used by EWT (~ 2.39 eV)
 
 # ================================================================
@@ -60,8 +60,8 @@ NEUTRINO_ENERGY = 3.8280e-19  # J, neutrino "seed" energy used by EWT (~ 2.39 eV
 # ================================================================
 ELECTRON_K = 10  # electron wave center count (dimensionless)
 ELECTRON_RADIUS = 2.8179403262e-15  # m, electron classical radius
-ELECTRON_ENERGY = 8.1871e-14  # J, electron rest energy (~ 0.511 MeV)
 ELECTRON_MASS = 9.1093837139e-31  # kg, electron mass from CODATA
+ELECTRON_ENERGY = 8.1871e-14  # J, electron rest energy (~ 0.511 MeV)
 ELECTRON_OUTER_SHELL = 2.138743820  # electron outer shell multiplier
 ELECTRON_ORBITAL_G = 0.9873318320  # electron orbital g-factor (gλ, dimensionless)
 ELECTRON_SPIN_G = 0.9826905018  # electron spin g-factor (gA, dimensionless)
@@ -74,8 +74,8 @@ ELECTRON_SPIN_G = 0.9826905018  # electron spin g-factor (gA, dimensionless)
 # ================================================================
 PROTON_K = 44  # proton wave center count (dimensionless)
 PROTON_RADIUS = 8.414e-16  # m, proton radius
-PROTON_ENERGY = 1.5033e-10  # J, CODATA proton rest energy (~ 938.272 MeV)
 PROTON_MASS = 1.67262192595e-27  # kg, proton mass from CODATA
+PROTON_ENERGY = 1.5033e-10  # J, CODATA proton rest energy (~ 938.272 MeV)
 PROTON_ORBITAL_G = 0.9898125300  # proton orbital g-factor (gp, dimensionless)
 
 NEUTRON_MASS = 1.67492749804e-27  # kg, neutron mass from CODATA 2022
@@ -134,6 +134,8 @@ IMPEDANCE_VACUUM = 376.730313412  # Ω, Z_0, characteristic impedance of vacuum,
 # Additional constants
 # ================================================================
 
+PI = np.pi  # π = 3.1415926535... (dimensionless)
+EULER_NUMBER = np.e  # e = 2.7182818284... (dimensionless)
 GOLDEN_RATIO = 1.6180339887  # φ = (1+sqrt(5))/2 (dimensionless)
 
 # DELTA : amplitude factor (dimensionless) — user-supplied per system/transition
