+++ openwave/xperiments/5_level1_wave_field/L1_launcher.py
@@ -19,7 +19,7 @@ import taichi as ti
 from openwave.common import colormap, constants
 from openwave._io import flux_mesh, render, video
 
-import openwave.spacetime.L1_field_based as data_grid
+import openwave.spacetime.L1_field_grid as data_grid
 import openwave.spacetime.L1_wave_engine as ewave
 import openwave.validations.wave_diagnostics as diagnostics
